+++ CHANGELOG.md
@@ -1142,3 +1142,13 @@ should use 4.0.1-alpha.0 for testing.
 #### web3-core
 
 -   Added rpc exception codes following eip-1474 as an experimental feature (if `useRpcCallSpecification` at `enableExperimentalFeatures` is `true`) (#5525)
+
+#### web3
+
+##### Removed
+
+-   Private static `_contracts:Contract[]` and static `setProvider` function was removed (#5792)
+
+##### Added
+
+-   `registeredSubscriptions` was added by default in web3 constructor (#5792)